seperate aligned feature extractor via inheritance
